Remove updatehistory field from NativeContractState and rename its base class which is now the NativeContractState class.
